requirements.txt
arize-phoenix 
gcsfs 
llama-index
PyGithub
gradio 
huggingface_hub
openai